moved flag toggle handler to domain, updated framework
