+++ src/FlagService.Domain/Features/Flags/Commands/FlagToggleCommandHandler.cs
@@ -3,10 +3,10 @@ using Framework2.Core.Extensions;
 using Framework2.Domain.Core.Handlers;
 using MediatR;
 
-namespace FlagService.Api.Features.Flags
+namespace FlagService.Domain.Features.Flags.Commands
 {
-    public class FlagToggleCommandHandler : 
-        FxCommandHandler<FlagToggleCommandRequest, FlagToggleCommandResponse>, 
+    public class FlagToggleCommandHandler :
+        FxCommandHandler<FlagToggleCommandRequest, FlagToggleCommandResponse>,
         IRequestHandler<FlagToggleCommandRequest, FlagToggleCommandResponse>
     {
         private readonly IFlagRepository _flagRepository;
