+++ src/FlagService.Domain/Features/Flags/Commands/FlagToggleCommandRequest.cs
@@ -1,7 +1,7 @@
 ﻿using Framework2.Domain.Core.Requests;
 using MediatR;
 
-namespace FlagService.Api.Features.Flags
+namespace FlagService.Domain.Features.Flags.Commands
 {
     public class FlagToggleCommandRequest : FxCommandRequest, IRequest<FlagToggleCommandResponse>
     {
