+++ src/FlagService.Domain/Features/Flags/Commands/FlagToggleCommandResponse.cs
@@ -1,7 +1,7 @@
 ﻿using FlagService.Domain.Aggregates;
 using Framework2.Domain.Core.Responses;
 
-namespace FlagService.Api.Features.Flags
+namespace FlagService.Domain.Features.Flags.Commands
 {
     public class FlagToggleCommandResponse : FxCommandResponse
     {
